requirements.txt
#
# Runtime dependencies.
#
suds-fc==0.4.1
zeep